% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiAGO2015}
\alias{ppiAGO2015}
\title{ppiAGO2015}
\format{A data frame with 9 columns and 101 rows:  
\describe{ 
\item{\code{score}}{PPI score}  
\item{\code{nl100}}{National poverty line (100)}  
\item{\code{nl150}}{National poverty line (150)}  
\item{\code{nl200}}{National poverty line (200)}  
\item{\code{half100}}{Poorest half below 100 national}  
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}  
\item{\code{ppp200}}{Below $2.00 per day purchasing power parity (2005)}  
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}  
\item{\code{ppp500}}{Below $5.00 per day purchasing power parity (2005)}  
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiAGO2015
}
\description{
Poverty Probability Index (PPI) lookup table for Angola
}
\keyword{datasets}
