% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiBOL2015}
\alias{ppiBOL2015}
\title{ppiBOL2015}
\format{A data frame with 10 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{Food poverty line}
\item{\code{nl150}}{National poverty line (100)}
\item{\code{nl200}}{National poverty line (150)}
\item{\code{half100}}{National poverty line (200)}
\item{\code{ppp125}}{Poorest half below 100 national}
\item{\code{ppp200}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.00 per day purchasing power parity (2005)}
\item{\code{ppp500}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{ppp844}}{Below $5.00 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiBOL2015
}
\description{
Poverty Probability Index (PPI) lookup table for Bolivia
}
\keyword{datasets}
