% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiCIV2018}
\alias{ppiCIV2018}
\title{ppiCIV2018}
\format{A data frame with 15 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nl150}}{National poverty line (150\%)}
\item{\code{nl200}}{National poverty line (200\%)}
\item{\code{ppp125}}{Below $1.00 per day purchasing power parity (2011)}
\item{\code{ppp250}}{Below $1.90 per day purchasing power parity (2011)}
\item{\code{ppp500}}{Below $3.20 per day purchasing power parity (2011)}
\item{\code{ppp100}}{Below $5.50 per day purchasing power parity (2011)}
\item{\code{ppp190}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp320}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{ppp550}}{Below $5.00 per day purchasing power parity (2005)}
\item{\code{percentile20}}{Below 20th percentile poverty line}
\item{\code{percentile40}}{Below 40th percentile poverty line}
\item{\code{percentile60}}{Below 60th percentile poverty line}
\item{\code{percentile80}}{Below 80th percentile poverty line}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiCIV2018
}
\description{
Poverty Probability Index (PPI) lookup table for Ivory Coast
}
\keyword{datasets}
