% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiCMR2013}
\alias{ppiCMR2013}
\title{ppiCMR2013}
\format{A data frame with 8 columns and 101 rows:  
\describe{ 
\item{\code{score}}{PPI score}  
\item{\code{nl100}}{National poverty line (100)}  
\item{\code{nl150}}{National poverty line (150)}  
\item{\code{nl200}}{National poverty line (200)}  
\item{\code{extreme}}{USAID extreme poverty}  
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}  
\item{\code{ppp200}}{Below $2.00 per day purchasing power parity (2005)}  
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}  
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiCMR2013
}
\description{
Poverty Probability Index (PPI) lookup table for Cameroon
}
\keyword{datasets}
