% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_ecuador.R
\docType{data}
\name{ppiECU2022}
\alias{ppiECU2022}
\title{Poverty Probability Index (PPI) lookup table for Ecuador for 2022}
\format{
A data frame with 20 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nl_extreme}}{National poverty line (extreme)}
\item{\code{nl150}}{National poverty line (150\%)}
\item{\code{nl200}}{National poverty line (200\%)}
\item{\code{ppp215}}{Below $2.15 per day purchasing power parity (2017)}
\item{\code{ppp365}}{Below $3.65 per day purchasing power parity (2017)}
\item{\code{ppp685}}{Below $6.85 per day purchasing power parity (2017)}
\item{\code{ppp100}}{Below $1.00 per day purchasing power parity (2011)}
\item{\code{ppp190}}{Below $1.90 per day purchasing power parity (2011)}
\item{\code{ppp320}}{Below $3.20 per day purchasing power parity (2011)}
\item{\code{ppp550}}{Below $5.50 per day purchasing power parity (2011)}
\item{\code{ppp800}}{Below $8.00 per day purchasing power parity (2011)}
\item{\code{ppp1100}}{Below $11.00 per day purchasing power parity (2011)}
\item{\code{ppp1500}}{Below $15.00 per day purchasing power parity (2011)}
\item{\code{ppp2170}}{Below $21.70 per day purchasing power parity (2011)}
\item{\code{percentile20}}{Below 20th percentile poverty line}
\item{\code{percentile40}}{Below 40th percentile poverty line}
\item{\code{percentile60}}{Below 60th percentile poverty line}
\item{\code{percentile80}}{Below 80th percentile poverty line}
}
}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiECU2022
}
\description{
Poverty Probability Index (PPI) lookup table for Ecuador for 2022
}
\examples{
  # Access Ecuador PPI table
  ppiECU2015

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiECU2015[ppiECU2015$score == ppiScore, ]

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiECU2015, score == ppiScore)

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the national
  # poverty line definition
  ppiScore <- 50
  ppiECU2015[ppiECU2015$score == ppiScore, "nl100"]

}
\keyword{datasets}
