% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiEGY2010}
\alias{ppiEGY2010}
\title{ppiEGY2010}
\format{A data frame with 8 columns and 101 rows:  
\describe{ 
\item{\code{score}}{PPI score}  
\item{\code{nu100}}{National upper poverty line (100)}  
\item{\code{nl100}}{National lower poverty line (100)}  
\item{\code{nlFood}}{Food poverty line}  
\item{\code{extreme}}{USAID extreme poverty}  
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}  
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}  
\item{\code{ppp375}}{Below $3.75 per day purchasing power parity (2005)}  
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiEGY2010
}
\description{
Poverty Probability Index (PPI) lookup table for Egypt
}
\keyword{datasets}
