% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiFJI2014}
\alias{ppiFJI2014}
\title{ppiFJI2014}
\format{A data frame with 8 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100)}
\item{\code{nl150}}{National poverty line (150)}
\item{\code{nl200}}{National poverty line (200)}
\item{\code{median}}{Poorest half below 100 national}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp200}}{Below $2.00 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiFJI2014
}
\description{
Poverty Probability Index (PPI) lookup table for Fiji
}
\keyword{datasets}
