% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_honduras.R
\docType{data}
\name{ppiHND2023}
\alias{ppiHND2023}
\title{Poverty Probability Index (PPI) lookup table for Honduras for 2023}
\format{
A data frame with 18 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nl_extreme}}{National poverty line (extreme)}
\item{\code{ppp100}}{Below $1.00 per day purchasing power parity (2011)}
\item{\code{ppp190}}{Below $1.90 per day purchasing power parity (2011)}
\item{\code{ppp320}}{Below $3.20 per day purchasing power parity (2011)}
\item{\code{ppp550}}{Below $5.50 per day purchasing power parity (2011)}
\item{\code{ppp800}}{Below $8.00 per day purchasing power parity (2011)}
\item{\code{ppp1100}}{Below $11.00 per day purchasing power parity (2011)}
\item{\code{ppp1500}}{Below $15.00 per day purchasing power parity (2011)}
\item{\code{ppp2170}}{Below $21.70 per day purchasing power parity (2011)}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{ppp500}}{Below $5.00 per day purchasing power parity (2005)}
\item{\code{percentile20}}{Below 20th percentile poverty line}
\item{\code{percentile40}}{Below 40th percentile poverty line}
\item{\code{percentile60}}{Below 60th percentile poverty line}
\item{\code{percentile80}}{Below 80th percentile poverty line}
}
}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiHND2023
}
\description{
Poverty Probability Index (PPI) lookup table for Honduras for 2023
}
\examples{
  # Access Honduras PPI table
  ppiHND2023

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiHND2023[ppiHND2023$score == ppiScore, ]

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiHND2023, score == ppiScore)

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the USAID
  # extreme poverty definition
  ppiScore <- 50
  ppiHND2023[ppiHND2023$score == ppiScore, "nl_extreme"]

}
\keyword{datasets}
