% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_indonesia.R
\docType{data}
\name{ppiIDN2012}
\alias{ppiIDN2012}
\title{Poverty Probability Index (PPI) lookup table for Indonesia using legacy
poverty definitions}
\format{
A data frame with 4 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
}
}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiIDN2012
}
\description{
Poverty Probability Index (PPI) lookup table for Indonesia using legacy
poverty definitions
}
\examples{
  # Access Indonesia PPI table
  ppiIDN2012

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiIDN2012[ppiIDN2012$score == ppiScore, ]

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiIDN2012, score == ppiScore)

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the national
  # poverty line definition
  ppiScore <- 50
  ppiIDN2012[ppiIDN2012$score == ppiScore, "nl100"]

}
\keyword{datasets}
