% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_indonesia.R
\docType{data}
\name{ppiIDN2012_a}
\alias{ppiIDN2012_a}
\title{Poverty Probability Index (PPI) lookup table for Indonesia using new
poverty definitions}
\format{
A data frame with 9 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nl150}}{National poverty line (150\%)}
\item{\code{nl200}}{National poverty line (200\%)}
\item{\code{extreme}}{USAID extreme poverty}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{ppp190}}{Below $1.90 per day purchasing power parity (2011)}
\item{\code{ppp310}}{Below $3.10 per day purchasing power parity (2011)}
}
}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiIDN2012_a
}
\description{
Poverty Probability Index (PPI) lookup table for Indonesia using new
poverty definitions
}
\examples{
  # Access Indonesia PPI table
  ppiIDN2012_a

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiIDN2012_a[ppiIDN2012_a$score == ppiScore, ]

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiIDN2012_a, score == ppiScore)

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the USAID
  # extreme poverty definition
  ppiScore <- 50
  ppiIDN2012_a[ppiIDN2012_a$score == ppiScore, "nl100"]

}
\keyword{datasets}
