% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiIND2016_r62}
\alias{ppiIND2016_r62}
\title{ppiIND2016_r62}
\format{A data frame with 7 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{saxena}}{National saxena}
\item{\code{ppp108}}{Below $1.08 per day purchasing power parity (1993)}
\item{\code{ppp81}}{Below $0.81 per day purchasing power parity (1993)}
\item{\code{ppp135}}{Below $1.35 per day purchasing power parity (1993)}
\item{\code{ppp162}}{Below $1.62 per day purchasing power parity (1993)}
\item{\code{ppp216}}{Below $2.16 per day purchasing power parity (1993)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiIND2016_r62
}
\description{
Poverty Probability Index (PPI) lookup table for India
}
\keyword{datasets}
