% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiIND2016_r66}
\alias{ppiIND2016_r66}
\title{ppiIND2016_r66}
\format{A data frame with 8 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{tendulkar}}{National tendulkar}
\item{\code{tendulkar100}}{National tendulkar (100)}
\item{\code{tendulkar150}}{National tendulkar (150)}
\item{\code{tendulkar200}}{National tendulkar (200)}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp188}}{Below $1.88 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiIND2016_r66
}
\description{
Poverty Probability Index (PPI) lookup table for India
}
\keyword{datasets}
