% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMatrixBRA}
\alias{ppiMatrixBRA}
\title{ppiMatrixBRA}
\format{A data frame with 10 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{belowHalfWage}}{Below the half minimum wage line}
\item{\code{belowQtrWage}}{Below the quarter minimum wage line}
\item{\code{belowOneWage}}{Below the one minimum wage line}
\item{\code{belowTwoWage}}{Below the two minimum wage line}
\item{\code{extreme}}{USAID extreme poverty}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{ppp375}}{Below $3.75 per day purchasing power parity (2005)}
\item{\code{ppp500}}{Below $5.00 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiMatrixBRA
}
\description{
Poverty Probability Index (PPI) lookup table for Brazil
}
\keyword{datasets}
