% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMatrixKHM}
\alias{ppiMatrixKHM}
\title{ppiMatrixKHM}
\format{A data frame with 6 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nl150}}{National poverty line (150\%)}
\item{\code{nl200}}{National poverty line (200\%)}
\item{\code{ppp125}}{Below $1.25 per day purchasing power poverty (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power poverty (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiMatrixKHM
}
\description{
Poverty Probability Index (PPI) lookup table for Cambodia
}
\keyword{datasets}
