% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMatrixMMR}
\alias{ppiMatrixMMR}
\title{ppiMatrixMMR}
\format{A data frame with 9 columns and 101 rows:
\describe{
\item{\code{X1}}{PPI score}
\item{\code{X2}}{Food poverty line}
\item{\code{X3}}{National poverty line (100\%)}
\item{\code{X4}}{National poverty line (150\%)}
\item{\code{X5}}{National poverty line (200\%)}
\item{\code{X6}}{USAID extreme poverty}
\item{\code{X7}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{X8}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{score}}{NA}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiMatrixMMR
}
\description{
Poverty Probability Index (PPI) lookup table for Myanmar
}
\keyword{datasets}
