% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMatrixROU}
\alias{ppiMatrixROU}
\alias{ppiMatrixROU}
\title{ppiMatrixROU}
\format{A data frame with 9 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100)}
\item{\code{nl150}}{National poverty line (150)}
\item{\code{nl200}}{National poverty line (200)}
\item{\code{extreme}}{Poorest half below 100 national}
\item{\code{ppp250}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp375}}{Below $2.00 per day purchasing power parity (2005)}
\item{\code{ppp500}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{laeken}}{Below $5.00 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}

\url{www.povertyindex.org}
}
\usage{
ppiMatrixROU

ppiMatrixROU
}
\description{
Poverty Probability Index (PPI) lookup table for Romania

Poverty Probability Index (PPI) lookup table for Romania
}
\keyword{datasets}
