% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMatrixRUS}
\alias{ppiMatrixRUS}
\title{ppiMatrixRUS}
\format{A data frame with 4 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{extreme}}{USAID extreme poverty}
\item{\code{ppp625}}{Below $6.25 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiMatrixRUS
}
\description{
Poverty Probability Index (PPI) lookup table for Russia
}
\keyword{datasets}
