% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiPHL2014}
\alias{ppiPHL2014}
\title{ppiPHL2014}
\format{A data frame with 6 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{ppp500}}{Below $5.00 per day purchasing power parity (2005)}
\item{\code{ppp432}}{Below $4.32 per day purchasing power parity (1993)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiPHL2014
}
\description{
Poverty Probability Index (PPI) lookup table for Philippines using legacy
poverty definitions
}
\keyword{datasets}
