% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiSLE2011}
\alias{ppiSLE2011}
\title{ppiSLE2011}
\format{A data frame with 8 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nlFood}}{Food poverty line}
\item{\code{nl75}}{National poverty line (75\%)}
\item{\code{nl150}}{National poverty line (150\%)}
\item{\code{extreme}}{USAID extreme poverty}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiSLE2011
}
\description{
Poverty Probability Index (PPI) lookup table for Sierra Leone
}
\keyword{datasets}
