% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ppitables.R
\docType{package}
\name{ppitables}
\alias{ppitables-package}
\alias{ppitables}
\title{Poverty Probability Index (PPI) Lookup Tables}
\description{
The Poverty Probability Index (PPI) is a poverty measurement tool for
organizations and businesses with a mission to serve the poor. The PPI is
statistically-sound, yet simple to use: the answers to 10 questions about a
household’s characteristics and asset ownership are scored to compute the
likelihood that the household is living below the poverty line – or above by
only a narrow margin.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katilingban/ppitables}
  \item \url{https://katilingban.io/ppitables/}
  \item Report bugs at \url{https://github.com/katilingban/ppitables/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernestgmd@gmail.com} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
