% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiVNM2009}
\alias{ppiVNM2009}
\title{ppiVNM2009}
\format{A data frame with 8 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nlFood}}{Food poverty line}
\item{\code{extreme}}{USAID extreme poverty line}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp175}}{Below $1.75 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2005)}
\item{\code{molisa}}{MOLISA poverty line}
}}
\source{
\url{www.povertyindex.org}
}
\usage{
ppiVNM2009
}
\description{
Poverty Probability Index (PPI) lookup table for Vietnam
}
\keyword{datasets}
