% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiCIV2013}
\alias{ppiCIV2013}
\title{Poverty Probability Index (PPI) lookup table for Ivory Coast}
\format{A data frame with 9 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100\%)}
\item{\code{nl150}}{National poverty line (150\%)}
\item{\code{nl200}}{National poverty line (200\%)}
\item{\code{extreme}}{USAID extreme poverty}
\item{\code{ppp125}}{Below $1.25 per day purchasing power parity (2005)}
\item{\code{ppp200}}{Below $2.00 per day purchasing power parity (2005)}
\item{\code{ppp250}}{Below $2.50 per day purchasing power parity (2011)}
\item{\code{ppp800}}{Below $8.00 per day purchasing power parity (2011)}
}}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiCIV2013
}
\description{
Poverty Probability Index (PPI) lookup table for Ivory Coast
}
\examples{
  # Access Ivory Coast PPI table
  ppiCIV2013

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiCIV2013[ppiCIV2013$score == ppiScore, ]

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiCIV2013, score == ppiScore)

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the USAID
  # extreme poverty definition
  ppiScore <- 50
  ppiCIV2013[ppiCIV2013$score == ppiScore, "extreme"]

}
\keyword{datasets}
