% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data.R
\docType{data}
\name{ppiIND2016_r62}
\alias{ppiIND2016_r62}
\title{Poverty Probability Index (PPI) lookup table for India using r62 poverty
definitions}
\format{
A data frame with 7 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{saxena}}{National saxena}
\item{\code{ppp108}}{Below $1.08 per day purchasing power parity (1993)}
\item{\code{ppp81}}{Below $0.81 per day purchasing power parity (1993)}
\item{\code{ppp135}}{Below $1.35 per day purchasing power parity (1993)}
\item{\code{ppp162}}{Below $1.62 per day purchasing power parity (1993)}
\item{\code{ppp216}}{Below $2.16 per day purchasing power parity (1993)}
}
}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiIND2016_r62
}
\description{
Poverty Probability Index (PPI) lookup table for India using r62 poverty
definitions
}
\examples{
  # Access India PPI table
  ppiIND2016_r62

  # Given a specific PPI score (from 0 - 100), get the row of poverty
  # probabilities from PPI table it corresponds to
  ppiScore <- 50
  ppiIND2016_r62[ppiIND2016_r62$score == ppiScore, ]

  # Use subset() function to get the row of poverty probabilities corresponding
  # to specific PPI score
  ppiScore <- 50
  subset(ppiIND2016_r62, score == ppiScore)

  # Given a specific PPI score (from 0 - 100), get a poverty probability
  # based on a specific poverty definition. In this example, the national
  # saxena poverty definition
  ppiScore <- 50
  ppiIND2016_r62[ppiIND2016_r62$score == ppiScore, "saxena"]

}
\keyword{datasets}
