% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppiMMR2019}
\alias{ppiMMR2019}
\title{Poverty Probability Index (PPI) lookup table for Myanmar}
\format{A data frame with 20 columns and 101 rows:
\describe{
\item{\code{score}}{PPI score}
\item{\code{nl100}}{National poverty line (100)}
\item{\code{extreme}}{National poverty line (150)}
\item{\code{nl150}}{National poverty line (200)}
\item{\code{nl200}}{Below $1.90 per day purchasing power parity (2011)}
\item{\code{ppp100}}{Below $3.20 per day purchasing power parity (2011)}
\item{\code{ppp190}}{Below $5.50 per day purchasing power parity (2011)}
\item{\code{ppp320}}{Below $8.00 per day purchasing power parity (2011)}
\item{\code{ppp550}}{Below $11.00 per day purchasing power parity (2011)}
\item{\code{ppp800}}{Below $15.00 per day purchasing power parity (2011)}
\item{\code{ppp1100}}{Below $21.70 per day purchasing power parity (2011)}
\item{\code{ppp1500}}{Below 20th percentile poverty line}
\item{\code{ppp2170}}{Below 40th percentile poverty line}
\item{\code{ppp125}}{Below 50th percentile poverty line}
\item{\code{ppp250}}{Below 60th percentile poverty line}
\item{\code{ppp500}}{Below 80th percentile poverty line}
\item{\code{percentile20}}{NA}
\item{\code{percentile40}}{NA}
\item{\code{percentile60}}{NA}
\item{\code{percentile80}}{NA}
}}
\source{
\url{https://www.povertyindex.org}
}
\usage{
ppiMMR2019
}
\description{
Poverty Probability Index (PPI) lookup table for Myanmar
}
\keyword{datasets}
